Card system v1.3
Added different card types as scripts for an easier creation and management.
