+++ Assets/Resources/Cards/w00_Sword.asset
@@ -9,15 +9,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e66455158c8eb654395272cecf6256de, type: 3}
+  m_Script: {fileID: 11500000, guid: 290df8450f704e74ea217b4aa1defdf2, type: 3}
   m_Name: w00_Sword
   m_EditorClassIdentifier: 
   name: Sword
-  description: Basic sword
+  description: A short sword
   artwork: {fileID: 0}
-  type: 0
-  effect: -1
-  effectValue: -1
   cost: 1
   weaponId: 0
-  attackDamage: -1
+  attackList: []
