+++ Assets/Scripts/CardScripts/CardData Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6004692fd1dea7f4491e13945da5ebd8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
