+++ Assets/Scripts/CardScripts/CardData Scripts/Attack.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Attack : CardData
+{
+    public int[] effect;
+    public int[] effectValue;
+    public int[] damage;
+}
