+++ Assets/Scripts/CardScripts/CardData Scripts/Attack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d6ccf39c6279f6a47958cf71147685b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
