+++ Assets/Scripts/CardScripts/CardData Scripts/Object.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 580270e59629f9f408e871ca53a21ba8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
