+++ Assets/Scripts/CardScripts/CardData Scripts/Weapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 290df8450f704e74ea217b4aa1defdf2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
